# Speed up model evaluation for some use cases