While working on modeling with @Cadair, we found a few places where some small changes can result in faster model evaluation.

* When evaluating 1D models, we can avoid calling ``np.broadcast_shapes`` to check if the inputs are broadcastable (the function is surprisingly slow for what it does)
* When parameters are scalars, we can likewise avoid some ``np.broadcast_shapes`` calls
* We don't have to always use ``units=True`` when calling ``_param_sets``, we can save a little time by not doing it for unitless models

This is actually quite hard to benchmark precisely as the times are very small anyway, but it looks like the speedup is on the order of 30-40% for evaluating ``Gaussian1D`` with scalars or small arrays. Benchmark code:

```python
import time
import numpy as np
from astropy.modeling.models import Gaussian1D

from pyinstrument import Profiler

g = Gaussian1D()

CASES = [
    ('Scalar', 2.0, 10000),
    ('Small array', np.linspace(-10, 10, 100), 10000),
    ('Medium array', np.linspace(-10, 10, 10_000), 1000),
    ('Large array', np.linspace(-10, 10, 1_000_000), 100)
]

# Scalar

for title, x, n in CASES:

    times = []

    for superiter in range(10):

        start = time.time()
        for iter in range(n):
            g(x)
        end = time.time()

        times.append((end - start) * 1e6 / n)

    print(f'{title:20s}: {np.median(times):8.3f} +/- {np.std(times)/np.sqrt(len(times)):8.3f}μs')
```

Before this PR:

```
Scalar              :   46.431 +/-    0.292μs
Small array         :   54.525 +/-    0.790μs
Medium array        :  140.993 +/-    1.386μs
Large array         : 9688.755 +/-  373.445μs
```

After this PR:

```
Scalar              :   29.580 +/-    0.429μs
Small array         :   33.545 +/-    0.433μs
Medium array        :   96.084 +/-    0.379μs
Large array         : 8678.083 +/-   84.160μs
```

This is 36%, 38%, 31%, and 10% faster (the improvement is not as significant for big arrays as expected, as for these the model evaluation dominates).

The asv benchmarks locally also show improvements, though the results are a bit noiser.

- [ ] By checking this box, the PR author has requested that maintainers do **NOT** use the "Squash and Merge" button. Maintainers should respect this when possible; however, the final decision is at the discretion of the maintainer that merges the PR.