@astrofrog 2024-07-08
I mean we could always just rename the changelog entry later if we wanted?