@pllim 2024-07-11
Not the increase I was expecting. Does this modeling boost not benchmarked over at benchmark repo?

```
| Change   | Before [f5126c76]    | After [5933a7d6]    |   Ratio | Benchmark (Parameter)          |
|----------|----------------------|---------------------|---------|--------------------------------|
| -        | 125±80μs             | 42.8±0.7μs          |    0.34 | units.time_quantity_times_unit |

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE INCREASED.
```