@astrofrog 2024-07-11
It is benchmarked but the improvement is on the order of 10% and the asv CI job has a threshold of 30%